Fixed qMin argument type for ARM
On ARM qReal is float and not double so we need to convert it first

+++ clients/studio/plugins/ThymioVPL/EventActionsSet.cpp
@@ -1206,7 +1206,7 @@ namespace Aseba { namespace ThymioVPL
 			qreal h,s,v,a;
 			color.getHsvF(&h, &s, &v, &a);
 			s *= Style::blockDropAreaSaturationFactor;
-			v = qMin(1.0, v*Style::blockDropAreaValueFactor);
+			v = qMin(1.0, (double)(v*Style::blockDropAreaValueFactor));
 			color.setHsvF(h,s,v,a);
 			//color.setAlpha(130);
 			painter->setPen(QPen(color, borderWidth, Qt::DotLine, Qt::RoundCap, Qt::RoundJoin));
